projects/detection/Signature_Verification_Portal/requirements.txt
numpy
pandas
streamlit
pillow
opencv-python
tensorflow
scikit-learn
scikit-image
fastdtw
db-sqlite3
